fix: resolve clippy 1.95 lints so the CI gate passes
Ported from claude/elastic-hopper-b4cc90 (565f94d). That branch's
cargo fmt commit is already covered by 276b964, so only the net lint
fixes are applied here. `Option::is_none_or` is avoided to keep the
Rust 1.75 MSRV.

+++ src/ui/panes.rs
@@ -43,8 +43,8 @@ fn collect_layout(area: Rect, tree: &PaneTree, out: &mut Vec<(PaneId, Rect)>) {
             let chunks = Layout::default()
                 .direction(direction)
                 .constraints([
-                    Constraint::Percentage(first_pct as u16),
-                    Constraint::Percentage(second_pct as u16),
+                    Constraint::Percentage(first_pct),
+                    Constraint::Percentage(second_pct),
                 ])
                 .split(area);
             collect_layout(chunks[0], first, out);
